Remove gettext macros in configure.ac, bug 650150
Using both intltool and gettext macros in configure.ac leads to problems
when autoreconf is run, as gettext is run by autopoint, which leads to
intltool rejecting the po/Makefile.in.in installed by gettext. Removing
the two gettext macro calls from configure.ac fixes the problem. This is
an acknowledged problem in the intltool documentation:

https://bugs.launchpad.net/intltool/+bug/784191