Use g_get_system_data_dirs() to find data files
In Vinagre.Dirs.get_package_data_file(), use g_get_system_data_dirs() to
loop through the data directories and return the path to the first data
file that exists. Check the path that was set when Vinagre was compiled
first, followed by those obtained from the system. In this way, it is
sufficient to set XDG_DATA_DIRS to make Vinagre relocatable, solving bug
653722.