Rework the first-run keyboard shortcut message
The first-run dialog was not suitable for a variety of reasons, such as
using complex language and not providing the ability to modify the
keyboard shortcut setting from within the dialog. Therefore, a
GtkInfoBar was used instead, the text was trimmed down to the useful
essentials and a button to enable the shortcuts directly was added, as
well as a button linking to the help. Fixes bug 609828.