Remove preferences dialog, bug 614774
The preferences dialog only had two options, neither of which were
particularly relevant in a dialog as opposed to a menu. The option to
always show tabs was removed from the UI, but can still be found in
GSettings. The option to enable keyboard shortcuts was moved to the View
menu.