Fix build on Cygwin, bug 665368
Cygwin uses gio-unix, not gio-windows.

Cygwin has posix_openpt but not getpt, which is nonstandard. Either will
work in pty_open.c, but in vinagre-ssh.c, USE_PTY is determined by
HAVE_GRANTPT, which is true on Cygwin. Using grantpt as the determining
factor across the board prevents any chance of error.

Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>