VinagreTab: don't crash on screenshot
In the save dialog, if the user changes the extension by clicking
on the filter, and he is in 'Recent Used' folder, vinagre crashes.

This is because gtk_file_chooser_get_filename() would return NULL
if we are in 'Recent Used' folder.

Workaround this issue in vinagre 3.8. A proper fix will reach master
(3.10) which depends on a newer gtk+ version.

https://bugzilla.gnome.org/show_bug.cgi?id=703185