VinagreTab: don't crash on screenshot
In the save dialog, if the user changes the extension by clicking
on the filter, and he is in 'Recent Used' folder, vinagre crashes.

This is because gtk_file_chooser_get_filename() would return NULL
if we are in 'Recent Used' folder.

Use the new gtk+ function gtk_file_chooser_get_current_name(), which
returns whatever is in the filename entry.

https://bugzilla.gnome.org/show_bug.cgi?id=703185