Fix credential access with libsecret, bug 704744
When storing attributes in libsecret, string attributes are not
permitted to have NULL values. The existing workaround for this was
wrong, as it removed the ability to store passwords for username-less
connections.  Fix this by terminating the attribute list with a NULL
attribute, thus ending the list, if there is no username.